Ydb stable 23-1-19
x-stable-origin-commit: c5d5a396e89d0a72e0267a55e93d8404d4fb54fe

+++ ydb/core/ymq/actor/queue_leader.h
@@ -90,7 +90,7 @@ private:
     void AnswerFailed(TSqsEvents::TEvGetConfiguration::TPtr& ev);
     void AskQueueAttributes();
     void OnQueueAttributes(const TSqsEvents::TEvExecuted::TRecord& ev);
-    void MarkInflyReloading(ui64 shard, size_t invalidatedCount, const TString& invalidationReason);
+    void MarkInflyReloading(ui64 shard, i64 invalidatedCount, const TString& invalidationReason);
     void StartLoadingInfly();
     void StartLoadingInfly(ui64 shard, bool afterFailure = false);
     void OnInflyLoaded(ui64 shard, const TSqsEvents::TEvExecuted::TRecord& reply);
